feat: [WIP] use proof keys to ensure requests come from trusted source
Basic code handling is in the proofkeys package. The idea is to
implement it as a middleware and hook it after the wopicontext
middleware.
Note that we need a way to switch this middleware off in case there are
problems with the verification, even though it should be active by
default.